fix: CsvBulkWriter flushes stream on every element added
Why:
Jackson writeValue flushes the underlying stream/writer by default. This behaviour
can be disabled by disabling the FLUSH_PASSED_TO_STREAM feature. However this breaks
the functionality of the bulk writer because Jackson internally has it's own buffering
components that we need to flushed. If you disable the flush, the CSV bytes are never
written to the underlying stream.

As a workaround we can wrap the output stream in an implementation that discards
flush calls, to prevent Jackson flushing the underlying stream (the CsvBulkWriter
will flush the underlying stream).